MySQL connector install and Spark Ports (#1)
* Install MySql connector
Looks like the MySql JAR is not installed when the Ambari server node is built. This fixes that missing file issue. Refer to https://community.hortonworks.com/questions/186090/mysql-connector-javajar-due-to-http-error-http-err.html for details

* Corrected yum install for MySql driver
Had forgotten the -y parameter

* More ports to expose
Added 18080 (Spark History Server port) and 4040 (Spark application default port) to the EXPOSE directive

* Spark port mappings

Had forgotten to add the port mappings for 4040 and 18080 to this script; fixed now.